requirements.txt
requests
pandas
openpyxl
crewai
fastapi
flask
openai
langchain-openai
python-dotenv
